\name{GeneralizedHyperbolic-internal}
\alias{ghypMomBNS}
\alias{ghypCreateParameters.R}
\alias{gigCreateParameters.R}

\title{Internal Generalized Hyperbolic Functions}
\description{
  Internal functions for the package GeneralizedHyperbolic
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
