\name{GeneralizedHyperbolic-internal}
\alias{ghypMomBNS}
\alias{ghypCreateParameters}
\alias{gigCreateParameters}
\alias{nigCreateParameters}
\alias{hyperbCreateParameters}
\alias{ghypStandPars}
\alias{hyperbStandPars}
\alias{kappaLambda}
\alias{alphaStand}
\alias{betaStand}
\alias{dsghyp}
\alias{dshyperb}
\alias{hyperbFitStand}
\alias{hyperbFitStandStart}
\alias{ghypScale}


\title{Internal Generalized Hyperbolic Functions}
\description{
  Internal functions for the package GeneralizedHyperbolic
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
